fix/time module

+++ bot/features/study_time/study_time.py
@@ -8,7 +8,7 @@ from bot import bot
 from discord import app_commands
 from bot.features.guild_event.on_member_join import on_member_join
 from models import Users, UserStudySection, UserDailyStudyTime
-from other_modules.time_modules import vn_now
+from other_modules.time_modules import Now
 
 updating_members = []
 
@@ -19,43 +19,45 @@ async def on_voice_state_update(
     member_before: discord.VoiceState,
     member_after: discord.VoiceState,
 ):
-    while member.id in updating_members:
-        await asyncio.sleep(1)
+    print("start study count")
+    if member.id == 880359404036317215:
+        if not member_before.channel and member_after.channel:
+            while member.id in updating_members:
+                await asyncio.sleep(1)
+            updating_members.append(member.id)
+            print(member.name, member.id, "in")
+            try:
+                user_study_section = await UserStudySection.find_one(
+                    UserStudySection.user.discord_id == str(member.id),
+                    fetch_links=True,
+                )
+                if user_study_section:
+                    user_study_section.start_study_time = Now().now
+                    await user_study_section.save()
+                else:
+                    await UserStudySection(
+                        user=await Users.find_one(Users.discord_id == str(member.id))
+                    ).insert()
 
-    if not member_before.channel and member_after.channel:
-        updating_members.append(member.id)
-        print(member.name, member.id, "in")
-        try:
-            user_study_section = await UserStudySection.find_one(
-                UserStudySection.user.discord_id == str(member.id),
-                fetch_links=True,
-            )
-            if user_study_section:
-                user_study_section.start_study_time = vn_now()
-                await user_study_section.save()
-            else:
+            except ValidationError as e:
+                print(e)
+                await on_member_join(member=member)
                 await UserStudySection(
                     user=await Users.find_one(Users.discord_id == str(member.id))
                 ).insert()
 
-        except ValidationError as e:
-            print(e)
-            await on_member_join(member=member)
-            await UserStudySection(
-                user=await Users.find_one(Users.discord_id == str(member.id))
-            ).insert()
-
-    elif member_before.channel and not member_after.channel:
-        print(member.name, member.id, "out")
-        user_study_section = await UserStudySection.find_one(
-            UserStudySection.user.discord_id == str(member.id), fetch_links=True
-        )
-        try:
-            await user_study_section.delete()
-        except AttributeError as e:
-            print(e)
-        if member.id in updating_members:
-            updating_members.remove(member.id)
+        elif member_before.channel and not member_after.channel:
+            print(member.name, member.id, "out")
+            user_study_section = await UserStudySection.find_one(
+                UserStudySection.user.discord_id == str(member.id), fetch_links=True
+            )
+            try:
+                await user_study_section.delete()
+            except AttributeError as e:
+                print(e)
+            if member.id in updating_members:
+                updating_members.remove(member.id)
+                print("not in updating member anymore")
 
 
 @bot.tree.command(name="study_time", description="Xem tổng thời gian học")
